Progress further with block validation tests

+++ lighthouse/state/common/maps.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 /// Maps a (slot, shard_id) to attestation_indices.
-pub type AttesterMap = HashMap<(u64, u64), Vec<usize>>;
+pub type AttesterMap = HashMap<(u64, u16), Vec<usize>>;
 
 /// Maps a slot to a block proposer.
 pub type ProposerMap = HashMap<u64, usize>;
